requirements/base.txt
# Bleeding edge Django
django==1.7.7

# Forms
django-braces==1.4.0
django-crispy-forms==1.4.0
django-floppyforms==1.3.0

# Models
django-model-utils==2.2

# Images
Pillow==2.7.0

# For user registration, either via email or social
# Well-built with regular release cycles!
django-allauth==0.18.0

# For the persistence stores
psycopg2==2.5.4

# Unicode slugification
unicode-slugify==0.1.1
django-autoslug==1.7.2

# Useful things
django-avatar==2.0
django-taggit==0.12.2
djangorestframework==3.1.1
Markdown==2.5.2
django-filter==0.9.1

# Time zones support
pytz==2014.10

#For LDAP Access / Login
django-auth-ldap==1.2.4
django-ldap-sync==0.1.1
